sorting and updating README
